+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- ICollisionable.CanCollide function
+
+### Changed
+- Check if can collide before Interactor enter in collision with ICollecable instance
+
 ## [2.1.0] - 2026-03-21
 ### Changed
 - Update package dependencies to use 1M Bits Horde